+++ tests/tests_unit/test_api/test_datapoints.py
@@ -1,6 +1,5 @@
 import json
 import math
-import re
 from contextlib import contextmanager
 from datetime import datetime
 from random import choice, random